feat: add metrics and move to infrastructure project
