+++ src/Conways.GameOfLife.API/Conways.GameOfLife.API.csproj
@@ -19,12 +19,6 @@
         <PrivateAssets>all</PrivateAssets>
         <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
       </PackageReference>
-      <PackageReference Include="OpenTelemetry" Version="1.9.0" />
-      <PackageReference Include="OpenTelemetry.Exporter.Console" Version="1.9.0" />
-      <PackageReference Include="OpenTelemetry.Exporter.OpenTelemetryProtocol" Version="1.9.0" />
-      <PackageReference Include="OpenTelemetry.Extensions.Hosting" Version="1.9.0" />
-      <PackageReference Include="OpenTelemetry.Instrumentation.AspNetCore" Version="1.9.0" />
-      <PackageReference Include="OpenTelemetry.Instrumentation.EntityFrameworkCore" Version="1.0.0-beta.12" />
       <PackageReference Include="Swashbuckle.AspNetCore" Version="6.6.2" />
     </ItemGroup>
